Move hooks to single utils folder
