+++ frontend/src/pages/AdvancedQueryBuilder/AdvancedQueryBuilder.tsx
@@ -3,7 +3,8 @@ import { Col, Row } from 'react-bootstrap';
 import { RouteComponentProps } from 'react-router-dom';
 import { Dimmer, Loader } from 'semantic-ui-react';
 import { OperationTabContainer } from '../../components/OperationTabContainer';
-import { useOperation, useSources } from '../../hooks';
+import { OperationMap } from '../../types/operations';
+import { useOperation, useSources } from '../../utils/hooks';
 
 type RouterParams = {
   id?: string;
@@ -12,7 +13,9 @@ type QueryBuilderProps = RouteComponentProps<RouterParams>;
 
 const AdvancedQueryBuilder: FunctionComponent<QueryBuilderProps> = (props) => {
   const { id: operationID } = props.match.params;
-  const { loading, operation } = useOperation(operationID ? parseInt(operationID) : undefined);
+  const { loading, operation } = useOperation<OperationMap>(
+    operationID ? parseInt(operationID) : undefined,
+  );
   const sources = useSources({ limit: 200, offset: 0 });
 
   return (
@@ -22,9 +25,7 @@ const AdvancedQueryBuilder: FunctionComponent<QueryBuilderProps> = (props) => {
           <Dimmer active={loading || !sources.count()} inverted>
             <Loader content="Loading" />
           </Dimmer>
-          {!loading && sources.count() ? (
-            <OperationTabContainer sources={sources.toJS()} operation={operation} />
-          ) : null}
+          {!loading && sources.count() ? <OperationTabContainer operation={operation} /> : null}
         </React.Fragment>
       </Col>
     </Row>
